[enh] Add Chrome/Chromium support

+++ backup_framapad_from_firefox.py
@@ -11,7 +11,7 @@ OS_TYPE = platform.system()
 HOME_DIR = os.getenv('HOME')
 BACKUP_DIR = HOME_DIR + '/FramapadBackup/'
 
-def retrieve_urls(profile):
+def retrieve_firefox_urls(profile):
     conn = sqlite3.connect(profile + 'places.sqlite')
 
     urls = conn.execute('''SELECT DISTINCT url
@@ -23,12 +23,25 @@ def retrieve_urls(profile):
 
     return urls
 
+def retrieve_chrome_urls(profile):
+    conn = sqlite3.connect(profile + 'History')
+
+    urls = conn.execute('''SELECT DISTINCT url
+                            FROM urls
+                            WHERE url
+                            LIKE "%lite%framapad.org/p/%"
+                            AND url NOT LIKE "%/timeslider%"
+                            AND url NOT LIKE "%/";''')
+
+    return urls
+
 def save_text(url):
     content = urlopen(url + '/export/txt')
     path = BACKUP_DIR + url.split('/')[-1]
 
-    with open(path + '.txt', 'w') as f:
-        f.write(content.read())
+    if not os.path.exists(path):
+        with open(path + '.txt', 'w') as f:
+            f.write(content.read())
 
 
 if __name__ == "__main__":
@@ -37,18 +50,35 @@ if __name__ == "__main__":
         os.mkdir(BACKUP_DIR)
 
     if OS_TYPE == "Linux":
-        profiles = glob(HOME_DIR + "/.mozilla/firefox/*.default/")
+        firefox_profiles = glob(HOME_DIR + "/.mozilla/firefox/*.default/")
+        chrome_profiles = glob(HOME_DIR + "/.config/google-chrome/Default/")
+        chromium_profiles = glob(HOME_DIR + "/.config/chromium/Default/")
+
     elif OS_TYPE == "Windows":
-        profiles = glob(os.getenv('APPDATA') + "\Mozilla\Firefox\Profiles\*.default\\")
+        firefox_profiles = glob(os.getenv('APPDATA') + "\Mozilla\Firefox\Profiles\*.default\\")
+
+        if platform.release() == "XP":
+            chrome_profiles = glob("C:\Documents and Settings\{}\Local Settings\Application Data\Google\Chrome\User Data\Default\\".format(os.getenv("USERNAME")))
+            chromium_profiles = glob("C:\Documents and Settings\{}\Local Settings\Application Data\Chromium\User Data\Default\\".format(os.getenv("USERNAME")))
+        else:
+            chrome_profiles = glob("C:\Users\{}\AppData\Local\Google\Chrome\User Data\Default\\".format(os.getenv("USERNAME")))
+            chromium_profiles = glob("C:\Users\{}\AppData\Local\Chromium\User Data\Default\\".format(os.getenv("USERNAME")))
     else:
-        profiles = glob(HOME_DIR + "/Library/Mozilla/.mozilla/firefox/*.default/")
+        firefox_profiles = glob(HOME_DIR + "/Library/Mozilla/.mozilla/firefox/*.default/")
+        chrome_profiles = glob(HOME_DIR + "/Library/Application Support/Google/Chrome/Default/")
+        chromium_profiles = glob(HOME_DIR + "/Library/Application Support/Chromium/Default/")
 
-    for profile in profiles:
-        urls = retrieve_urls(profile)
+    urls = []
 
-        for url in urls:
-            print("Téléchargement de {}".format(url[0]))
-            save_text(url[0])
+    for profile in firefox_profiles:
+        urls += retrieve_firefox_urls(profile)
+
+    for profile in chrome_profiles + chromium_profiles:
+        urls += retrieve_chrome_urls(profile)
+
+    for url in urls:
+        print("Téléchargement de {}".format(url[0]))
+        save_text(url[0])
 
     print("Tous les pads trouvés ont été téléchargés ici : {}".format(BACKUP_DIR))
